Made code more typesafe
For some reason, tsc was throwing an error that ctx was of type any. Had to throw in some eslint exceptions while I moved on. Came back to remove the eslint exceptions and the tsc error was no longer there.
Also finally figured out a good way to make post.date typesafe in getPost.ts