Created a higher order function to check for admin role
Was using useEffect to redirect users when they tried to access an admin only page without the admin role. Created a higher order function to use with getServerSideProps insteads. This means I only need to put the code on the index page of the admin route and the entire route is protected.